+++ src/main/java/until/the/eternity/common/request/PageRequestDto.java
@@ -1,4 +1,4 @@
-package until.the.eternity.common.dto;
+package until.the.eternity.common.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Max;
@@ -9,7 +9,7 @@ import org.springframework.data.domain.Sort;
 
 @Schema(description = "페이지 요청 파라미터")
 public record PageRequestDto(
-        @Schema(description = "요청할 페이지 번호 (0부터 시작)", example = "1") @Min(1) Integer page,
+        @Schema(description = "요청할 페이지 번호 (1부터 시작)", example = "1") @Min(1) Integer page,
         @Schema(description = "페이지당 항목 수", example = "20") @Min(1) @Max(100) Integer size,
         @Schema(description = "정렬 필드 (예: createdAt)", example = "createdAt") String sortBy,
         @Schema(description = "정렬 방향 (asc or desc)", example = "desc") String direction) {
@@ -20,7 +20,7 @@ public record PageRequestDto(
     private static final String DEFAULT_DIRECTION = "desc";
 
     public Pageable toPageable() {
-        int resolvedPage = this.page != null ? this.page : DEFAULT_PAGE;
+        int resolvedPage = this.page != null ? this.page - 1 : DEFAULT_PAGE;
         int resolvedSize = this.size != null ? this.size : DEFAULT_SIZE;
         String resolvedSortBy = this.sortBy != null ? this.sortBy : DEFAULT_SORT_BY;
         Sort.Direction resolvedDirection = parseDirection(this.direction);
